// types.ts
export interface Question {
  id: number;
  text: string;
  options: string[];
  correctAnswer: number; // Index
  topic: string;
  difficulty: 'Easy' | 'Medium' | 'Hard';
  explanation?: string;
  isRepeated?: boolean; // New: Is this a frequently repeated question?
  frequency?: number;   // New: How many times appeared in past years
  subject?: string;     // Added for section segmentation
}
// ...
export interface Topic {
  id: string;
  name: string;
  subject: string;
}
// ...
export interface ExamPatternSection {
    subject: string;
    questionCount: number;
}

export interface ExamPattern {
    examId: string;
    durationMinutes: number;
    totalQuestions: number;
    sections: ExamPatternSection[];
}